fix: #289 eliminate bottom and right margins from label sheets
also update dependency to use openpdf
