--- lib/com/lowagie/legacy/itext/4.2.1/_remote.repositories
@@ -1,4 +1,0 @@
-#NOTE: This is an Aether internal implementation file, its format can be changed without prior notice.
-#Wed Sep 25 10:07:49 BRT 2019
-itext-4.2.1.pom>=
-itext-4.2.1.jar>=
