--- lib/com/lowagie/legacy/itext/4.2.1/m2e-lastUpdated.properties
@@ -1,7 +1,0 @@
-#Mon Apr 30 22:37:55 BRT 2018
-k-int.com|http\://developer.k-int.com/maven2/|sources=1525138072022
-central|http\://central.maven.org/maven2/|javadoc=1525138675402
-k-int.com|http\://developer.k-int.com/maven2/|javadoc=1525138675402
-central|http\://central.maven.org/maven2/|sources=1525138072022
-clojars.org|http\://clojars.org/repo|javadoc=1525138675402
-clojars.org|http\://clojars.org/repo|sources=1525138072022
