--- lib/com/lowagie/legacy/itext/maven-metadata-local.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>com.lowagie.legacy</groupId>
-  <artifactId>itext</artifactId>
-  <versioning>
-    <release>4.2.1</release>
-    <versions>
-      <version>4.2.1</version>
-    </versions>
-    <lastUpdated>20190925130749</lastUpdated>
-  </versioning>
-</metadata>
